% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{conf2tex}
\alias{conf2tex}
\title{print a confusion matrix as tex}
\usage{
conf2tex(conf, caption = "", label = "")
}
\arguments{
\item{conf}{the confusion matrix, a table}

\item{caption}{the caption for the table, defaults to ""}

\item{label}{the label for the table, defaults to ""}
}
\description{
\code{conf2tex} prints a 2x2 confusion matrix as a latex table preservering the margin
}
