% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distcalc}
\alias{distcalc}
\title{Calculate distances between character and numeric variables}
\usage{
distcalc(
  dat,
  character_variables = c("mlast", "mfirst", "wfirst", "minitials", "winitials",
    "mprof"),
  numeric_variables = c("year")
)
}
\arguments{
\item{dat}{a dataset. \code{distcalc} expects this dataset 
to follow the naming conventions of the \code{candidates} 
function: the variables from the original datasets have the 
same name and are distinguished using \code{_from} and 
\code{_to} suffixes.}

\item{character_variables}{The names of the character variables, without the suffixes.  Set to a length zero vector if no variables are to be used (\code{c()}).}

\item{numeric_variables}{The names of the numeric variables, without the suffixes.  Set to a length zero vector if no variables are to be used (\code{c()}).}
}
\value{
the dataset with the necessary distances to predict links. Note that reassignment of the data.table is not necessary. The original dataset is modified in place
}
\description{
\code{distcalc} calculates string distances and numeric distances
}
\examples{
d1 = data.table::data.table(mlast = c("jong", "smid"), persid = c(1:2))
d2 = data.table::data.table(mlast = c("jongh", "jong", "smit"), persid = c(1:3))
d1d2cnd = candidates(d1, d2)
distcalc(d1d2cnd, character_variables = "mlast", numeric_variables = c())
d1d2cnd

}
