% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{expand_index}
\alias{expand_index}
\title{Supplement an existing linking index with new links.}
\usage{
expand_index(dat)
}
\arguments{
\item{dat}{a dataset containing a variable index which is to be supplemented, and a variable candidate_index for the indexes which are to supplement it}
}
\description{
\code{exand_index}supplements an existing linking index using a 
new set of proposed links.
}
\examples{
dat = data.table(
     index = c(1, 1, 1, 2, 2, NA, 2, NA, NA, NA, 4, 4, NA),
     index_candidate = c(5, 5, NA, 6, 6, 6, 6, 7, 7, NA, 8, 8, 8))
expand_index(dat)
}
