% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gk}
\alias{gk}
\title{Two dimensional Gaussian kernel}
\usage{
gk(x1, x2, sigma = 1)
}
\arguments{
\item{x1}{the numbers in x1 that should be compared to those in x2}

\item{x2}{the numbers in x2 that should be compared to those in x1}

\item{sigma}{parameter controlling the spread of the gaussian curve. The 
higher, the more tolerant we are are of differences. Defaults to 1.}
}
\description{
\code{gk} calculates the Gaussian similarity as the parallel distances between the elements of two vectors
}
\details{
The Gaussian similarity is calculated as \eqn{d(x_1, x_2) = e ^{ - \frac{\sqrt{(x_1 - x_2) ^ 2}}{\sigma^2}}}
}
