% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{initials}
\alias{initials}
\title{Create initials from a string of names}
\usage{
initials(strings)
}
\arguments{
\item{strings}{string containing names, seperated by spaces or periods (or both). Vectorised.}
}
\value{
Initials in the form JF, so no spaces and no periods.
}
\description{
\code{initials} takes a string containing names, seperated either by spaces of periods, and returns the initials.
}
\examples{
strings = c("Auke B.", "Arnold Christoffer", "Arend X. Mark", "Albert X Johannes", "Albert X.Y. Johannes", "J. F.")
initials(strings)
}
