% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Manually label data}
\usage{
label(
  dat_from,
  dat_to,
  persid_from,
  persid_to,
  blockvariable,
  blocktype,
  N,
  path,
  ...
)
}
\arguments{
\item{dat_from}{a data.table}

\item{dat_to}{a data.table}

\item{persid_from}{string identifying the person id variable}

\item{persid_to}{string identifying the person id variable}

\item{blockvariable}{string identifying the blocking variable}

\item{N}{the number of unique observations of the blocking varaible 
to be labelled, defaults to 500}

\item{...}{passed to candidates for customised blocking}
}
\value{
A list containing candidate pairs to be labelled
}
\description{
\code{label} wraps a sampling and the candidates function to make manual labelling of training data easier
}
\details{
label takes a random sample from dat_from, gathers candidates from dat_to
and presents them to the user to select the match or tell that there is no match

The labelling session is interactive, and the user is presented with a 
choice between
\itemize{
     \item{Persid}{One of the numbers of persid_to}
     \item{None}
}
At some point a "Back" option might be added

After selecting there is an annotation step, that can be done 
 \itemize{
     \item{Cancel}
     \item{Sure}
     \item{Maybe}
     \item{Doubtful}
     \item{Ambiguous}
}
}
\examples{
d1 = data.table::data.table(mlast = c("jong", "smid"), mfirst = c("Jan", "Jan"), wfirst = NA, wlast = NA, settlerchildren = NA, persid = c(1:2))
d2 = data.table::data.table(mlast = c("jongh", "jong", "smit"), mfirst = c("Jan", "Dirk", "Johan"), wlast = NA, wfirst = NA, settlerchildren = NA, persid = c(1:3))
label(d1, d2, "persid", "persid", "mlast", "bigram distance", 2)

}
