% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{len_longest_word}
\alias{len_longest_word}
\title{character length of a the longest word in a string}
\usage{
len_longest_word(str)
}
\arguments{
\item{str}{character vector to search in}
}
\description{
\code{len_longest_word} calculates the character length of a the longest
 word in a string.
}
