% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predict_links}
\alias{predict_links}
\title{Predict links}
\usage{
predict_links(dat_candidates, id_from, id_to, minimum_confidence = 0.5,
  modstring = c("m_rf_baptisms_full", "m_rf_baptisms_sparse"),
  linktype = c("one:one", "many:one"),
  some_measure_of_other_close_matches = "not_implemented")
}
\arguments{
\item{id_from}{string giving the identifier variable in the candidates dataset for the observations from the \code{_from} dataset.}

\item{id_to}{string giving the identifier variable in the candidates dataset for the observations from the \code{_to} dataset.}

\item{minimum_confidence}{the minimum confidence level (vote share) to return a link. Defaults to 0.5.}

\item{modstring}{String giving the name of one of the pretrained models, either m_rf_baptisms_sparse or m_rf_baptisms_full. See details below.}

\item{dat_candidates.}{A dataset with link candidates, created using \code{make_candidates}, and with distances calculated using \code{distcalc}. \code{predict_links} expects this dataset to follow the naming conventions of the model you're using (detailed below).}
}
\value{
The candidates dataset filtered down to only the best links for each record in the original \code{_from} dataset.
Also some details about one:one and many:one
}
\description{
\code{predict links} takes a dataset of candidate links and distances, to predict links and select the best ones.
}
\details{
uses datatable to handle potentially large numer of candidates
no need to assign to a new object
if you do this, the original dataset is also modified
maybe I should fix this?

The following models can be used 
\itemize{
 \item \code{m_rf_baptisms_sparse} a model linking parents in baptism records to marriage records, based on minimal information: male surname (mlast), male first name (mfirst), female first name (wfirst, female surname not used because it would typically not be reported in the baptism records), and year of marriage/baptism (year).
 \item \code{m_rf_baptisms_full} a model linking parents in baptism and marriage records, using additional information: initials, profession, and soundex distances of the names. Performance is not much better than the sparse model.
}
}
