% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rand_strings_like}
\alias{rand_strings_like}
\title{Generate random strings resembling old strings}
\usage{
rand_strings_like(strings_like, n_samples = 1)
}
\arguments{
\item{strings_like}{a vector of strings that the random strings should resemble}
}
\description{
\code{rand_strings_like} creates new strings with the same distributions of letters and string lengths as a vector of supplied strings. Needs to be at least length two.
}
