% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rm_diacretics}
\alias{rm_diacretics}
\title{Remove diacretics from letters}
\usage{
rm_diacretics(strings)
}
\arguments{
\item{strings}{string containing names, seperated by spaces or periods (or both). Vectorised.}
}
\value{
Initials in the form JF, so no spaces and no periods.
}
\description{
\code{rm_diacretics} replaces letters with diacretics (like "é") with letters without diacretics (like "e"). \code{iconv(..., to = ASCII//TRANSLIT)} might also work, but fails for me sometimes.
}
\details{
Removing diacretics might be useful if they are being used inconsistently in the data being linked. Differences in diacretics count in the calculation of string distances.

The list of diacretics is currently far from complete, only what I encountered in the baptism and marriage records I looked at.
}
\examples{
rm_diacretics(strings = "éå")
iconv("éå", "UTF-8", "ASCII//TRANSLIT") # bit unpredicatble for me

}
