% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{split_prefixes}
\alias{split_prefixes}
\title{Split out prefixes from surname strings}
\usage{
split_prefixes(strings, more_prefixes = c())
}
\arguments{
\item{strings}{a vector of strings containing surnames}

\item{more_prefixes}{a vector of additional prefixes that are not
currently in the list already}
}
\description{
\code{split_prefixes()} uses a list of common South African prefixes
 to split out prefixes from surnames
}
\examples{
split_prefixes(strings = c("VAN DER MOLEN", "DUMOULIN", "DU MOULIN"))
split_prefixes(strings = c("Van der Molen", "DUMOULIN", "DU MOULIN"))

}
