% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{stringdist_closest}
\alias{stringdist_closest}
\title{Calculate closets string distance to another string in a vector}
\usage{
stringdist_closest(string)
}
\description{
Calculate closets string distance to another string in a vector
}
