% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{uniformise_string}
\alias{uniformise_string}
\title{Uniformise strings}
\usage{
uniformise_string(string, maxdist = 0.2, quiet = FALSE)
}
\description{
This is not very effective because of the use of stringdistmat
but at least it does it on the unique names
}
